+++ app/src/main/java/ru/kuchanov/scpquiz/model/api/NwQuizTransaction.kt
@@ -10,6 +10,6 @@ data class NwQuizTransaction(
         var userId: Long? = null,
         var quizId: Long? = null,
         val quizTransactionType: TransactionType,
-        val coinsAmount: Long? = null
+        val coinsAmount: Int? = null
 )
